Extract host-service bootstrap from provider_build.go and unify duplicate IndexedDB host-service wiring so the PR is net deletive vs main. Host-service assembly, relay env wiring, public registration, cache/S3/IndexedDB binding servers, and the scoped IndexedDB build chain now live in host_service_bootstrap.go, with shared public relay URL parsing beside the runtime placement plan in plugin_runtime_plan.go. This amend collapses newIndexedDBHostService, indexedDBHostService, and externalCredentialsIndexedDBHostService into one registerIndexedDBServer path, dedupes binding-map construction, and consolidates public host-service registration helpers, with no behavior change.
