+++ gestaltd/internal/bootstrap/provider_build.go
@@ -8,7 +8,6 @@ import (
 	"io"
 	"log/slog"
 	"maps"
-	"net"
 	"net/url"
 	"path/filepath"
 	"slices"
@@ -20,21 +19,14 @@ import (
 	coreagent "github.com/valon-technologies/gestalt/server/core/agent"
 	corecache "github.com/valon-technologies/gestalt/server/core/cache"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
-	"github.com/valon-technologies/gestalt/server/core/indexeddb"
-	s3store "github.com/valon-technologies/gestalt/server/core/s3"
 	coreworkflow "github.com/valon-technologies/gestalt/server/core/workflow"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	proto "github.com/valon-technologies/gestalt/server/internal/gen/v1"
 	"github.com/valon-technologies/gestalt/server/internal/invocationconfig"
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
 	agentservice "github.com/valon-technologies/gestalt/server/services/agents"
-	authorizationservice "github.com/valon-technologies/gestalt/server/services/authorization"
-	cacheservice "github.com/valon-technologies/gestalt/server/services/cache"
 	"github.com/valon-technologies/gestalt/server/services/egress"
 	"github.com/valon-technologies/gestalt/server/services/egressproxy"
-	externalcredentialsservice "github.com/valon-technologies/gestalt/server/services/externalcredentials"
 	"github.com/valon-technologies/gestalt/server/services/identity/principal"
-	indexeddbservice "github.com/valon-technologies/gestalt/server/services/indexeddb"
 	"github.com/valon-technologies/gestalt/server/services/invocation"
 	"github.com/valon-technologies/gestalt/server/services/observability/metricutil"
 	plugininvokerservice "github.com/valon-technologies/gestalt/server/services/plugininvoker"
@@ -52,10 +44,7 @@ import (
 	"github.com/valon-technologies/gestalt/server/services/providerdrivers/componentprovider"
 	"github.com/valon-technologies/gestalt/server/services/runtimehost"
 	"github.com/valon-technologies/gestalt/server/services/runtimehost/pluginruntime"
-	"github.com/valon-technologies/gestalt/server/services/s3"
-	workflowservice "github.com/valon-technologies/gestalt/server/services/workflows"
 	"github.com/valon-technologies/gestalt/server/services/workflows/workflowmanager"
-	"google.golang.org/grpc"
 	"gopkg.in/yaml.v3"
 )
 
@@ -1545,10 +1534,7 @@ func newLocalPluginRuntime(runtimeProviderName string, deps Deps) pluginruntime.
 }
 
 const (
-	pluginRuntimeHostServiceRelayTokenTTL = 30 * 24 * time.Hour
-	pluginRuntimeEgressProxyTokenTTL      = 30 * 24 * time.Hour
-	hostServiceTLSCAFileEnv               = "GESTALT_HOST_SERVICE_TLS_CA_FILE"
-	hostServiceTLSCAPEMEnv                = "GESTALT_HOST_SERVICE_TLS_CA_PEM"
+	pluginRuntimeEgressProxyTokenTTL = 30 * 24 * time.Hour
 )
 
 type RuntimeEgressLaunchPlan struct {
@@ -1693,272 +1679,6 @@ func waitForPluginRuntimeSessionReady(ctx context.Context, runtimeProvider plugi
 	}
 }
 
-func buildPluginRuntimeHostServices(name string, entry *config.ProviderEntry, deps Deps) ([]runtimehost.HostService, *plugininvokerservice.InvocationTokenManager, func(), error) {
-	var (
-		hostServices []runtimehost.HostService
-		cleanup      func()
-		invTokens    *plugininvokerservice.InvocationTokenManager
-	)
-	fail := func(err error) ([]runtimehost.HostService, *plugininvokerservice.InvocationTokenManager, func(), error) {
-		if cleanup != nil {
-			cleanup()
-			cleanup = nil
-		}
-		return nil, nil, nil, err
-	}
-
-	effectiveIndexedDB, err := config.ResolveEffectivePluginIndexedDB(name, entry, deps.SelectedIndexedDBName, deps.IndexedDBDefs)
-	if err != nil {
-		return fail(err)
-	}
-	if effectiveIndexedDB.Enabled {
-		services, indexedDBCleanup, err := buildIndexedDBHostServices(name, effectiveIndexedDB.ProviderName, effectiveIndexedDB, deps)
-		if err != nil {
-			return fail(err)
-		}
-		hostServices = append(hostServices, services...)
-		cleanup = chainCleanup(cleanup, indexedDBCleanup)
-	}
-	if len(entry.Cache) > 0 {
-		services, cacheCleanup, err := buildPluginCacheHostServices(name, entry, deps)
-		if err != nil {
-			return fail(err)
-		}
-		hostServices = append(hostServices, services...)
-		cleanup = chainCleanup(cleanup, cacheCleanup)
-	}
-	if len(entry.S3) > 0 {
-		services, err := buildPluginS3HostServices(name, entry, deps)
-		if err != nil {
-			return fail(err)
-		}
-		hostServices = append(hostServices, services...)
-	}
-	includeWorkflowProvider := deps.WorkflowManager != nil || (deps.WorkflowRuntime != nil && deps.WorkflowRuntime.HasConfiguredProviders())
-	includeAgentProvider := deps.AgentManager != nil || deps.AgentRuntime != nil
-	needInvocationTokens := len(entry.Invokes) > 0
-	if includeWorkflowProvider || includeAgentProvider {
-		needInvocationTokens = true
-	}
-	if needInvocationTokens {
-		invTokens, err = plugininvokerservice.NewInvocationTokenManager(deps.EncryptionKey)
-		if err != nil {
-			return fail(err)
-		}
-	}
-	if includeWorkflowProvider {
-		hostServices = append(hostServices, buildPluginWorkflowProviderHostService(name, deps, invTokens))
-	}
-	if includeAgentProvider {
-		hostServices = append(hostServices, buildPluginAgentProviderHostService(name, deps, invTokens))
-	}
-	if deps.Services != nil && !core.ExternalCredentialProviderMissing(deps.Services.ExternalCredentials) {
-		hostServices = append(hostServices, buildPluginExternalCredentialsHostService(deps.Services.ExternalCredentials))
-	}
-	if deps.AuthorizationProvider != nil && len(entry.EffectiveHTTPBindings()) > 0 {
-		hostServices = append(hostServices, buildPluginAuthorizationHostService(deps.AuthorizationProvider))
-	}
-	if len(entry.Invokes) > 0 {
-		hostServices = append(hostServices, buildPluginInvokerHostService(name, entry, deps, invTokens))
-	}
-	return hostServices, invTokens, cleanup, nil
-}
-
-func appendRuntimeLogHostService(hostServices []runtimehost.HostService, runtimeConfig config.EffectiveRuntimePlacement, deps Deps, runtimePlan RuntimePlacementPlan) []runtimehost.HostService {
-	if deps.Services == nil || deps.Services.RuntimeSessionLogs == nil || runtimePlan.Resolved.HostServiceAccess == RuntimeHostServiceAccessNone {
-		return hostServices
-	}
-	runtimeProviderName := runtimeSessionLogProviderName(runtimeConfig)
-	return append(hostServices, runtimehost.HostService{
-		Name:           "runtime_log_host",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.PluginRuntimeLogHost_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			runtimehost.RegisterRuntimeLogHostServer(srv, runtimeProviderName, deps.Services.RuntimeSessionLogs.AppendSessionLogs)
-		},
-	})
-}
-
-func runtimeSessionLogProviderName(runtimeConfig config.EffectiveRuntimePlacement) string {
-	if name := strings.TrimSpace(runtimeConfig.ProviderName); name != "" {
-		return name
-	}
-	return "local"
-}
-
-func withRuntimeSessionEnv(env map[string]string, sessionID string) map[string]string {
-	sessionID = strings.TrimSpace(sessionID)
-	if sessionID == "" {
-		return env
-	}
-	if env == nil {
-		env = map[string]string{}
-	}
-	env[runtimehost.DefaultRuntimeSessionIDEnv] = sessionID
-	return env
-}
-
-func withHostServiceTLSCAEnv(env map[string]string, deps Deps) map[string]string {
-	caPEM := strings.TrimSpace(deps.HostServiceTLSCAPEM)
-	caFile := strings.TrimSpace(deps.HostServiceTLSCAFile)
-	if caPEM == "" && caFile == "" {
-		return env
-	}
-	if env == nil {
-		env = map[string]string{}
-	}
-	if caPEM != "" {
-		env[hostServiceTLSCAPEMEnv] = caPEM
-	} else {
-		env[hostServiceTLSCAFileEnv] = caFile
-	}
-	return env
-}
-
-type hostServiceBindingDescriptor struct {
-	Name           string
-	MethodPrefixes []string
-}
-
-func hostServiceBindingDescriptorFromConfigured(hostService runtimehost.HostService) hostServiceBindingDescriptor {
-	return hostServiceBindingDescriptor{
-		Name:           strings.TrimSpace(hostService.Name),
-		MethodPrefixes: append([]string(nil), hostService.MethodPrefixes...),
-	}
-}
-
-func grpcMethodPrefix(serviceName string) string {
-	return "/" + serviceName + "/"
-}
-
-func hostServiceBindingDescriptorsFromConfigured(hostServices []runtimehost.HostService) []hostServiceBindingDescriptor {
-	if len(hostServices) == 0 {
-		return nil
-	}
-	out := make([]hostServiceBindingDescriptor, 0, len(hostServices))
-	for _, hostService := range hostServices {
-		out = append(out, hostServiceBindingDescriptorFromConfigured(hostService))
-	}
-	return out
-}
-
-func mergeHostedRuntimeHostServiceRelayEnv(providerName, sessionID string, hostServices []hostServiceBindingDescriptor, deps Deps) (map[string]string, string, error) {
-	if len(hostServices) == 0 {
-		return nil, "", nil
-	}
-	for _, hostService := range hostServices {
-		if strings.TrimSpace(hostService.Name) == "" || len(hostService.MethodPrefixes) == 0 {
-			return nil, "", fmt.Errorf("host service %q requires public host service relay support", hostService.Name)
-		}
-	}
-	serviceLabel := strings.ReplaceAll(strings.TrimSpace(hostServices[0].Name), "_", " ")
-	relayDialTarget, relayEnv, relayHost, ok, err := buildHostedRuntimePublicHostServiceRelay(
-		providerName,
-		sessionID,
-		hostServices[0],
-		deps,
-		serviceLabel,
-	)
-	if err != nil {
-		return nil, "", err
-	}
-	if !ok {
-		return nil, "", fmt.Errorf("provider %q requires server.baseURL and server.encryptionKey to relay host services through the public host service relay", providerName)
-	}
-	relayEnv[runtimehost.DefaultHostServiceSocketEnv] = relayDialTarget
-	return relayEnv, relayHost, nil
-}
-
-type runtimeHostServiceSessionVerifier struct {
-	providerName string
-	provider     pluginruntime.Provider
-}
-
-func (v runtimeHostServiceSessionVerifier) VerifyHostServiceSession(ctx context.Context, sessionID string) error {
-	sessionID = strings.TrimSpace(sessionID)
-	if sessionID == "" {
-		return fmt.Errorf("runtime session id is required")
-	}
-	if v.provider == nil {
-		return fmt.Errorf("plugin runtime provider is not configured")
-	}
-	session, err := v.provider.GetSession(ctx, pluginruntime.GetSessionRequest{SessionID: sessionID})
-	if err != nil {
-		return err
-	}
-	if session == nil {
-		return fmt.Errorf("plugin runtime session %q was not found", sessionID)
-	}
-	if expected := strings.TrimSpace(v.providerName); expected != "" {
-		if got := strings.TrimSpace(session.Metadata["provider_name"]); got != "" && got != expected {
-			return fmt.Errorf("plugin runtime session %q belongs to provider %q", sessionID, got)
-		}
-	}
-	if session.Lifecycle != nil && session.Lifecycle.ExpiresAt != nil {
-		expiresAt := session.Lifecycle.ExpiresAt.UTC()
-		if !time.Now().UTC().Before(expiresAt) {
-			return fmt.Errorf("plugin runtime session %q expired at %s", sessionID, expiresAt.Format(time.RFC3339Nano))
-		}
-	}
-	switch session.State {
-	case pluginruntime.SessionStatePending, pluginruntime.SessionStateReady, pluginruntime.SessionStateRunning:
-		return nil
-	default:
-		return fmt.Errorf("plugin runtime session %q is %s", sessionID, session.State)
-	}
-}
-
-func registerPublicRuntimeHostServices(providerName string, hostServices []runtimehost.HostService, deps Deps, runtimePlan RuntimePlacementPlan, runtimeProvider pluginruntime.Provider) (func(), error) {
-	if runtimePlan.Resolved.HostServiceAccess != RuntimeHostServiceAccessRelay || deps.PublicHostServices == nil {
-		return nil, nil
-	}
-	registerHostServices := publicRuntimeRegistryHostServices(hostServices)
-	if len(registerHostServices) == 0 {
-		return nil, nil
-	}
-	for _, hostService := range registerHostServices {
-		if strings.TrimSpace(hostService.Name) == "" {
-			return nil, fmt.Errorf("host service %q requires a service name for public relay", hostService.Name)
-		}
-	}
-	registration := deps.PublicHostServices.RegisterVerified(providerName, runtimeHostServiceSessionVerifier{
-		providerName: providerName,
-		provider:     runtimeProvider,
-	}, registerHostServices...)
-	return func() {
-		registration.Unregister()
-	}, nil
-}
-
-type workflowHostServiceSessionVerifier struct{}
-
-func (workflowHostServiceSessionVerifier) VerifyHostServiceSession(context.Context, string) error {
-	// Non-runtime workflow providers do not allocate runtime sessions; the
-	// signed relay token is scoped by provider, service, and method.
-	return nil
-}
-
-func registerPublicWorkflowHostServices(providerName string, hostServices []runtimehost.HostService, deps Deps) (func(), error) {
-	if deps.PublicHostServices == nil || !hostCanRelayPluginRuntimeHostServices(deps) {
-		return nil, nil
-	}
-	registerHostServices := make([]runtimehost.HostService, 0, len(hostServices))
-	for _, hostService := range hostServices {
-		if hostService.Register == nil {
-			continue
-		}
-		if strings.TrimSpace(hostService.Name) == "" {
-			return nil, fmt.Errorf("host service %q requires a service name for public relay", hostService.Name)
-		}
-		registerHostServices = append(registerHostServices, hostService)
-	}
-	if len(registerHostServices) == 0 {
-		return nil, nil
-	}
-	registration := deps.PublicHostServices.RegisterVerified(providerName, workflowHostServiceSessionVerifier{}, registerHostServices...)
-	return func() {
-		registration.Unregister()
-	}, nil
-}
 
 func buildHostedRuntimePublicEgressProxy(providerName, sessionID string, allowedHosts []string, defaultAction egress.PolicyAction, deps Deps) (map[string]string, error) {
 	baseURL, explicitRelayBaseURL := hostedRuntimeRelayBaseURL(deps)
@@ -1991,98 +1711,6 @@ func buildHostedRuntimePublicEgressProxy(providerName, sessionID string, allowed
 	}, nil
 }
 
-func publicRuntimeRegistryHostServices(hostServices []runtimehost.HostService) []runtimehost.HostService {
-	if len(hostServices) == 0 {
-		return nil
-	}
-	return append([]runtimehost.HostService(nil), hostServices...)
-}
-
-func buildHostedRuntimePublicHostServiceRelay(providerName, sessionID string, hostService hostServiceBindingDescriptor, deps Deps, serviceLabel string) (string, map[string]string, string, bool, error) {
-	baseURL, explicitRelayBaseURL := hostedRuntimeRelayBaseURL(deps)
-	if baseURL == "" || len(deps.EncryptionKey) == 0 {
-		return "", nil, "", false, nil
-	}
-	dialTarget, relayHost, err := pluginRuntimePublicRelayTarget(baseURL, explicitRelayBaseURL)
-	if err != nil {
-		return "", nil, "", false, err
-	}
-	tokenManager, err := runtimehost.NewHostServiceRelayTokenManager(deps.EncryptionKey)
-	if err != nil {
-		return "", nil, "", false, fmt.Errorf("init host service relay tokens: %w", err)
-	}
-	token, err := tokenManager.MintToken(runtimehost.HostServiceRelayTokenRequest{
-		PluginName: providerName,
-		SessionID:  sessionID,
-		Service:    "host_service",
-		// MethodPrefix "/" scopes the relay to the unified host-service surface.
-		// Per-RPC access is enforced by AllowsMethod and session verification, not
-		// by narrowing the token to individual gRPC service prefixes.
-		MethodPrefix: "/",
-		TTL:          pluginRuntimeHostServiceRelayTokenTTL,
-	})
-	if err != nil {
-		return "", nil, "", false, fmt.Errorf("mint %s host service relay token: %w", serviceLabel, err)
-	}
-	return dialTarget, map[string]string{
-		runtimehost.DefaultHostServiceTokenEnv: token,
-	}, relayHost, true, nil
-}
-
-func pluginRuntimePublicRelayTarget(baseURL string, allowInsecureHTTP bool) (string, string, error) {
-	parsed, host, err := pluginRuntimePublicProxyBaseURL(baseURL, allowInsecureHTTP)
-	if err != nil {
-		return "", "", err
-	}
-	port := parsed.Port()
-	if port == "" {
-		if strings.EqualFold(parsed.Scheme, "http") {
-			port = "80"
-		} else {
-			port = "443"
-		}
-	}
-	target := net.JoinHostPort(host, port)
-	switch strings.ToLower(parsed.Scheme) {
-	case "https":
-		return "tls://" + target, host, nil
-	case "http":
-		return "tcp://" + target, host, nil
-	default:
-		return "", "", fmt.Errorf("server.baseURL %q has unsupported public runtime relay scheme %q", baseURL, parsed.Scheme)
-	}
-}
-
-func pluginRuntimePublicProxyBaseURL(baseURL string, allowInsecureHTTP bool) (*url.URL, string, error) {
-	parsed, err := url.Parse(strings.TrimSpace(baseURL))
-	if err != nil {
-		return nil, "", fmt.Errorf("parse server.baseURL for public runtime relay: %w", err)
-	}
-	host := strings.TrimSpace(parsed.Hostname())
-	if host == "" {
-		return nil, "", fmt.Errorf("server.baseURL %q is missing a hostname", baseURL)
-	}
-	if path := strings.TrimSpace(parsed.EscapedPath()); path != "" && path != "/" {
-		return nil, "", fmt.Errorf("server.baseURL %q must not include a path for public runtime relay", baseURL)
-	}
-	if parsed.RawQuery != "" || parsed.Fragment != "" {
-		return nil, "", fmt.Errorf("server.baseURL %q must not include a query or fragment for public runtime relay", baseURL)
-	}
-	switch strings.ToLower(parsed.Scheme) {
-	case "https":
-	case "http":
-		if !allowInsecureHTTP && !isLoopbackAllowedHost(host) {
-			return nil, "", fmt.Errorf("server.baseURL %q must use https for public runtime relay unless it targets loopback", baseURL)
-		}
-	default:
-		return nil, "", fmt.Errorf("server.baseURL %q must use https for public runtime relay", baseURL)
-	}
-	parsed.Path = ""
-	parsed.RawPath = ""
-	parsed.RawQuery = ""
-	parsed.Fragment = ""
-	return parsed, host, nil
-}
 
 func appendAllowedHost(allowedHosts []string, host string) []string {
 	host = strings.TrimSpace(host)
@@ -2119,215 +1747,6 @@ func hostedAgentAllowedHosts(allowedHosts []string, runtimePlan RuntimePlacement
 	return out
 }
 
-func isLoopbackAllowedHost(host string) bool {
-	host = strings.TrimSpace(strings.Trim(host, "[]"))
-	if strings.EqualFold(host, "localhost") {
-		return true
-	}
-	ip := net.ParseIP(host)
-	return ip != nil && ip.IsLoopback()
-}
-
-func buildIndexedDBHostServices(serverLabel, metricsName string, effective config.EffectiveHostIndexedDBBinding, deps Deps) ([]runtimehost.HostService, func(), error) {
-	if deps.IndexedDBFactory == nil || len(deps.IndexedDBDefs) == 0 {
-		return nil, nil, fmt.Errorf("indexeddb host services are not available")
-	}
-
-	ds, err := buildHostScopedIndexedDB(metricsName, effective, deps)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	hostServices := []runtimehost.HostService{
-		newIndexedDBHostService(serverLabel, ds, effective.ObjectStores),
-	}
-	return hostServices, func() {
-		_ = closeIndexedDBs(ds)
-	}, nil
-}
-
-func newIndexedDBHostService(serverLabel string, ds indexeddb.IndexedDB, allowedStores []string) runtimehost.HostService {
-	opts := indexeddbservice.ServerOptions{AllowedStores: allowedStores}
-	return runtimehost.HostService{
-		Name:           "indexeddb",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.IndexedDB_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterIndexedDBServer(srv, indexeddbservice.NewServer(ds, serverLabel, opts))
-		},
-	}
-}
-
-func buildPluginCacheHostServices(pluginName string, entry *config.ProviderEntry, deps Deps) ([]runtimehost.HostService, func(), error) {
-	if deps.CacheFactory == nil || len(deps.CacheDefs) == 0 {
-		return nil, nil, fmt.Errorf("cache host services are not available")
-	}
-
-	bindings := make(map[string]corecache.Cache, len(entry.Cache))
-	boundCaches := make([]corecache.Cache, 0, len(entry.Cache))
-	for _, bindingName := range entry.Cache {
-		def, ok := deps.CacheDefs[bindingName]
-		if !ok || def == nil {
-			_ = closeCaches(boundCaches...)
-			return nil, nil, fmt.Errorf("cache %q is not available", bindingName)
-		}
-		value, err := buildCache(def, &FactoryRegistry{Cache: deps.CacheFactory})
-		if err != nil {
-			_ = closeCaches(boundCaches...)
-			return nil, nil, fmt.Errorf("cache %q: %w", bindingName, err)
-		}
-		boundCaches = append(boundCaches, value)
-		bindings[bindingName] = value
-	}
-	defaultBinding := ""
-	if len(boundCaches) == 1 {
-		defaultBinding = entry.Cache[0]
-	}
-	hostServices := []runtimehost.HostService{{
-		Name:           "cache",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.Cache_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterCacheServer(srv, registerCacheServer(bindings, defaultBinding, pluginName))
-		},
-	}}
-	return hostServices, func() {
-		_ = closeCaches(boundCaches...)
-	}, nil
-}
-
-func buildPluginS3HostServices(pluginName string, entry *config.ProviderEntry, deps Deps) ([]runtimehost.HostService, error) {
-	if len(deps.S3) == 0 {
-		return nil, fmt.Errorf("s3 host services are not available")
-	}
-
-	var accessURLs *s3.ObjectAccessURLManager
-	if len(deps.EncryptionKey) != 0 {
-		var err error
-		accessURLs, err = s3.NewObjectAccessURLManager(deps.EncryptionKey, deps.BaseURL)
-		if err != nil {
-			return nil, fmt.Errorf("s3 object access URLs: %w", err)
-		}
-	}
-
-	bindings := make(map[string]s3store.Client, len(entry.S3))
-	for _, binding := range entry.S3 {
-		client, ok := deps.S3[binding]
-		if !ok || client == nil {
-			return nil, fmt.Errorf("s3 %q is not available", binding)
-		}
-		bindings[binding] = client
-	}
-	defaultBinding := ""
-	if len(entry.S3) == 1 {
-		defaultBinding = entry.S3[0]
-	}
-	hostServices := []runtimehost.HostService{{
-		Name: "s3",
-		MethodPrefixes: []string{
-			grpcMethodPrefix(proto.S3_ServiceDesc.ServiceName),
-			grpcMethodPrefix(proto.S3ObjectAccess_ServiceDesc.ServiceName),
-		},
-		Register: func(srv *grpc.Server) {
-			s3Server, objectAccessServer := registerS3Servers(bindings, defaultBinding, pluginName, accessURLs)
-			proto.RegisterS3Server(srv, s3Server)
-			proto.RegisterS3ObjectAccessServer(srv, objectAccessServer)
-		},
-	}}
-	return hostServices, nil
-}
-
-func registerIndexedDBServer(bindings map[string]indexeddb.IndexedDB, defaultBinding, pluginName string, opts indexeddbservice.ServerOptions) proto.IndexedDBServer {
-	if len(bindings) == 1 {
-		for _, ds := range bindings {
-			return indexeddbservice.NewServer(ds, pluginName, opts)
-		}
-	}
-	return indexeddbservice.NewRoutingServer(bindings, defaultBinding, pluginName, opts)
-}
-
-func registerCacheServer(bindings map[string]corecache.Cache, defaultBinding, pluginName string) proto.CacheServer {
-	if len(bindings) == 1 {
-		for _, cache := range bindings {
-			return cacheservice.NewServer(cache, pluginName)
-		}
-	}
-	return cacheservice.NewRoutingServer(bindings, defaultBinding, pluginName)
-}
-
-func registerS3Servers(bindings map[string]s3store.Client, defaultBinding, pluginName string, accessURLs *s3.ObjectAccessURLManager) (proto.S3Server, proto.S3ObjectAccessServer) {
-	if len(bindings) == 1 {
-		for bindingName, client := range bindings {
-			return s3.NewServer(client, pluginName), s3.NewObjectAccessServer(accessURLs, pluginName, bindingName)
-		}
-	}
-	return s3.NewRoutingServers(bindings, defaultBinding, pluginName, accessURLs)
-}
-
-func buildPluginWorkflowProviderHostService(pluginName string, deps Deps, tokens *plugininvokerservice.InvocationTokenManager) runtimehost.HostService {
-	manager := deps.WorkflowManager
-	if manager == nil {
-		manager = unavailableWorkflowManager{}
-	}
-	return runtimehost.HostService{
-		Name:           "workflow_provider",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.WorkflowProvider_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterWorkflowProviderServer(srv, workflowservice.NewProviderServer(pluginName, manager, tokens))
-		},
-	}
-}
-
-func buildPluginAgentProviderHostService(pluginName string, deps Deps, tokens *plugininvokerservice.InvocationTokenManager) runtimehost.HostService {
-	manager := deps.AgentManager
-	if manager == nil {
-		manager = unavailableAgentManager{}
-	}
-	return runtimehost.HostService{
-		Name:           "agent_provider",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.AgentProvider_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterAgentProviderServer(srv, agentservice.NewProviderServer(pluginName, manager, tokens))
-		},
-	}
-}
-
-func buildPluginAuthorizationHostService(provider core.AuthorizationProvider) runtimehost.HostService {
-	return runtimehost.HostService{
-		Name:           "authorization",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.AuthorizationProvider_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterAuthorizationProviderServer(srv, authorizationservice.NewProviderServer(provider))
-		},
-	}
-}
-
-func buildPluginExternalCredentialsHostService(provider core.ExternalCredentialProvider) runtimehost.HostService {
-	return runtimehost.HostService{
-		Name:           "external_credentials",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.ExternalCredentialProvider_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterExternalCredentialProviderServer(srv, externalcredentialsservice.NewProviderServer(provider))
-		},
-	}
-}
-
-func buildPluginInvokerHostService(pluginName string, entry *config.ProviderEntry, deps Deps, tokens *plugininvokerservice.InvocationTokenManager) runtimehost.HostService {
-	invoker := deps.PluginInvoker
-	if invoker == nil {
-		invoker = unavailablePluginInvoker{}
-	}
-	return runtimehost.HostService{
-		Name:           "plugin_invoker",
-		MethodPrefixes: []string{grpcMethodPrefix(proto.PluginInvoker_ServiceDesc.ServiceName)},
-		Register: func(srv *grpc.Server) {
-			proto.RegisterPluginInvokerServer(srv, plugininvokerservice.NewServer(
-				pluginName,
-				invocationconfig.PluginInvocationDependencies(entry.Invokes),
-				invoker,
-				tokens,
-			))
-		},
-	}
-}
 
 type unavailablePluginInvoker struct{}
 
@@ -2503,138 +1922,6 @@ func (unavailableAgentManager) ResolveInteraction(context.Context, *principal.Pr
 	return nil, fmt.Errorf("agent manager is not available")
 }
 
-func buildHostScopedIndexedDB(metricsName string, effective config.EffectiveHostIndexedDBBinding, deps Deps) (indexeddb.IndexedDB, error) {
-	return buildScopedIndexedDB(scopedIndexedDBBuildOptions{
-		MetricsName:  metricsName,
-		ProviderName: effective.ProviderName,
-		DB:           effective.DB,
-	}, deps)
-}
-
-type scopedIndexedDBBuildOptions struct {
-	MetricsName  string
-	ProviderName string
-	DB           string
-}
-
-func buildScopedIndexedDB(opts scopedIndexedDBBuildOptions, deps Deps) (indexeddb.IndexedDB, error) {
-	def, ok := deps.IndexedDBDefs[opts.ProviderName]
-	if !ok || def == nil {
-		return nil, fmt.Errorf("indexeddb %q is not available", opts.ProviderName)
-	}
-	scopedDef, err := newScopedIndexedDBDef(def, scopedIndexedDBDefOptions{
-		DB: opts.DB,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("indexeddb %q: %w", opts.ProviderName, err)
-	}
-	ds, err := buildIndexedDB(scopedDef, &FactoryRegistry{IndexedDB: deps.IndexedDBFactory})
-	if err != nil {
-		return nil, fmt.Errorf("indexeddb %q: %w", opts.ProviderName, err)
-	}
-	return metricutil.InstrumentIndexedDB(ds, opts.MetricsName), nil
-}
-
-type scopedIndexedDBDefOptions struct {
-	DB string
-}
-
-func newScopedIndexedDBDef(entry *config.ProviderEntry, opts scopedIndexedDBDefOptions) (*config.ProviderEntry, error) {
-	if entry == nil {
-		return nil, fmt.Errorf("datastore provider is required")
-	}
-	cfg, err := config.NodeToMap(entry.Config)
-	if err != nil {
-		return nil, fmt.Errorf("decode config: %w", err)
-	}
-	if cfg == nil {
-		cfg = make(map[string]any)
-	}
-
-	switch {
-	case isRelationalIndexedDBEntry(entry):
-		if isSQLiteIndexedDBConfig(cfg) {
-			delete(cfg, "schema")
-			cfg["table_prefix"] = opts.DB + "_"
-			cfg["prefix"] = opts.DB + "_"
-		} else {
-			delete(cfg, "table_prefix")
-			delete(cfg, "prefix")
-			cfg["schema"] = opts.DB
-		}
-	case isMongoDBIndexedDBEntry(entry):
-		cfg["database"] = opts.DB
-	case isDynamoDBIndexedDBEntry(entry):
-		cfg["table"] = opts.DB
-	default:
-		return nil, fmt.Errorf("scoped indexeddb bindings require a provider with config-level namespace support")
-	}
-
-	configNode, err := mapToYAMLNode(cfg)
-	if err != nil {
-		return nil, fmt.Errorf("encode config: %w", err)
-	}
-
-	cloned := *entry
-	cloned.Config = configNode
-	return &cloned, nil
-}
-
-func isRelationalIndexedDBEntry(entry *config.ProviderEntry) bool {
-	return isIndexedDBProviderEntry(entry, "relationaldb")
-}
-
-func isMongoDBIndexedDBEntry(entry *config.ProviderEntry) bool {
-	return isIndexedDBProviderEntry(entry, "mongodb")
-}
-
-func isDynamoDBIndexedDBEntry(entry *config.ProviderEntry) bool {
-	return isIndexedDBProviderEntry(entry, "dynamodb")
-}
-
-func isIndexedDBProviderEntry(entry *config.ProviderEntry, providerName string) bool {
-	if entry == nil {
-		return false
-	}
-	providerPath := "/indexeddb/" + providerName
-	if entry.ResolvedManifest != nil {
-		source := strings.TrimSpace(filepath.ToSlash(entry.ResolvedManifest.Source))
-		return strings.HasSuffix(source, providerPath)
-	}
-	if metadataURL := strings.TrimSpace(entry.SourceMetadataURL()); metadataURL != "" {
-		parsed, err := url.Parse(metadataURL)
-		if err == nil {
-			path := filepath.ToSlash(parsed.Path)
-			return strings.Contains(path, providerPath+"/") && strings.HasSuffix(path, "/provider-release.yaml")
-		}
-	}
-	if path := strings.TrimSpace(entry.SourcePath()); path != "" {
-		path = filepath.ToSlash(path)
-		return strings.HasSuffix(path, providerPath) ||
-			strings.HasSuffix(path, "/"+providerName) ||
-			strings.HasSuffix(path, providerPath+"/manifest.yaml") ||
-			strings.HasSuffix(path, "/"+providerName+"/manifest.yaml")
-	}
-	return false
-}
-
-func isSQLiteIndexedDBConfig(cfg map[string]any) bool {
-	dsn, _ := cfg["dsn"].(string)
-	dsn = strings.TrimSpace(dsn)
-	if dsn == "" {
-		return false
-	}
-	switch {
-	case strings.HasPrefix(dsn, "postgres://"), strings.HasPrefix(dsn, "postgresql://"):
-		return false
-	case strings.HasPrefix(dsn, "mysql://"), strings.Contains(dsn, "@tcp("), strings.Contains(dsn, "@unix("):
-		return false
-	case strings.HasPrefix(dsn, "sqlserver://"):
-		return false
-	default:
-		return true
-	}
-}
 
 func mapToYAMLNode(value map[string]any) (yaml.Node, error) {
 	data, err := yaml.Marshal(value)
