+++ gestaltd/internal/server/host_service_public_test.go
@@ -10,20 +10,20 @@ import (
 	"google.golang.org/grpc"
 )
 
-func TestHostServiceHandlerReturnsVerifierError(t *testing.T) {
+func TestUnifiedHostServiceHandlerReturnsVerifierError(t *testing.T) {
 	t.Parallel()
 
 	registry := runtimehost.NewPublicHostServiceRegistry()
 	registry.RegisterVerified("support", rejectHostServiceSessionVerifier{}, testHostService())
 	s := &Server{publicHostServices: registry}
 
-	handler, err := s.hostServiceHandler(context.Background(), runtimehost.HostServiceRelayTarget{
+	handler, err := s.unifiedHostServiceHandler(context.Background(), runtimehost.HostServiceRelayTarget{
 		PluginName: "support",
 		SessionID:  "session-1",
 		Service:    "cache",
 	}, "/gestalt.provider.v1.Cache/Get")
 	if err == nil || !strings.Contains(err.Error(), `runtime session "session-1" is not active`) {
-		t.Fatalf("hostServiceHandler err = %v, want verifier failure", err)
+		t.Fatalf("unifiedHostServiceHandler err = %v, want verifier failure", err)
 	}
 	if handler != nil {
 		t.Fatalf("handler = %v, want no handler", handler)
@@ -37,25 +37,25 @@ func TestUnifiedHostServiceHandlerRoutesByGRPCMethod(t *testing.T) {
 	registry.RegisterVerified("support", allowHostServiceSessionVerifier{}, testHostService())
 	s := &Server{publicHostServices: registry}
 
-	handler, err := s.hostServiceHandler(context.Background(), runtimehost.HostServiceRelayTarget{
+	handler, err := s.unifiedHostServiceHandler(context.Background(), runtimehost.HostServiceRelayTarget{
 		PluginName: "support",
 		SessionID:  "session-1",
 		Service:    "host_service",
 	}, "/gestalt.provider.v1.Cache/Get")
 	if err != nil {
-		t.Fatalf("hostServiceHandler: %v", err)
+		t.Fatalf("unifiedHostServiceHandler: %v", err)
 	}
 	if handler == nil {
 		t.Fatal("handler = nil, want cache handler")
 	}
 
-	handler, err = s.hostServiceHandler(context.Background(), runtimehost.HostServiceRelayTarget{
+	handler, err = s.unifiedHostServiceHandler(context.Background(), runtimehost.HostServiceRelayTarget{
 		PluginName: "support",
 		SessionID:  "session-1",
 		Service:    "host_service",
 	}, "/gestalt.provider.v1.S3/ListObjects")
 	if err != nil {
-		t.Fatalf("hostServiceHandler unregistered method: %v", err)
+		t.Fatalf("unifiedHostServiceHandler unregistered method: %v", err)
 	}
 	if handler != nil {
 		t.Fatal("handler for unregistered method is non-nil")
